Prefix all refdes's with S1/ to prevent naming collisions when inserting into board layout.
Unfortunately PCB doesn't have the ability to do heirachical layout designs. This is more of a hack and a better way would be to get the refdes of the sub-schematic symbol and have it automatically prefixed to the refdes's in the interconnect PCB layout file. The prefixed PCB layouts could then be inserted into the board layout without any problems. This method only allows one interconnect block per board without some bodging and assumes the SRIC symbol has the refdes 'S1' in the board schematic.