Change power LED to green. Add Token In LED.
Token In LED will illuminate if the board is plugged in correctly. If the 'IN' port is plugged into another 'IN' port then it will not illuminate.
Needs prototyping as the high threshold for the Vpp pin of nC is 0.8Vdd and the high output of the RS485 receiver isn't 5V.