Remove test point footprints from 'in' and 'out' symbols
These were only added to give something to route the 'external' nets to in PCB. Now the PCB layout is done they are no longer needed and are actually a nuisance. Then appear when you first place the SRIC layout into a board but then disappear the next time you run gsch2pcb due to gEDA treating the 'in' and 'out' symbols specially, their footprints are only used when generating a PCB layout from the schematic directly, not when useing the schematic as a sub-schematic inside another.