Remove RS485 transceivers. Change main reg to 3.3V.
It should be a case of dropping in LVDS drivers/receivers in place of the
RS485 ones. We will no longer need a 5V rail so the main 800mA 5V reg can be
replaced with a 3.3V one.

Due to the massive voltage drop across it we will only be able to draw around
220mA from the 3.3V rail. If a board requires more current then it'll have to
provide its own regulation from the 12V rail.