Move token LED so it illuminates when the TI line is low.
The LVDS drivers idle high when there is no input connected and in this state
the token LED should not be on. By moving the LED between Vcc and the TI line
it will come on when the TI line goes low. This requires the polarity of the
token to be changed in the interconnect firmware.

Both the TI and TO lines are now active low.