Add Polyfuse/resistors to differential lines.
Add 4 polyfuses to the data and token out differential lines. These provide a
current limit through the TVS diodes when a fault is present. As the token in
lines only connect to an LVDS receiver they can be sufficiently protected with
resistors.

Due to the LVDS differential output drivers being current sources/sinks the
added resistance of the polyfuses should not affect the performance greatly.