Add values to token in differential line protection resistors.
The LVDS receiver inputs have a maximum input current of 20uA. This gives
a maximum voltage drop of 6.6mV across a 330ohm resistor which should not
affect the performance of the receiver significantly.

Assuming a fault voltage of 12V the 3V TVS diode will clamp at around 6V,
dropping 6V across the resistor. The fault current will be ~20mA. This gives
a power dissipation of 120mW in both the TVS diode and the resistor.