Add delays before and after sending a byte
The delay between bytes allows the receiver to re-synchronise if
it drops a byte, transmitting bytes constantly doesn't allow the
receiver to detect the start/end of a byte correctly.

The 20us delay is needed after sending the byte to allow it to be
fully transmitted before disabling the driver. A better way to
achieve this will be to wait until any data has finished being
transmitted before disabling the driver.